feat(core): Support combat state/time attributes and use long for time
Added handling for AttrCombatState and AttrCombatStateTime in MessageAnalyzer. Changed PlayerInfo.CombatStateTime from int to long and updated all related methods and interfaces to use long. Updated design-time and mock implementations to match the new signature. This enables accurate storage of player combat state and its timestamp.

+++ StarResonanceDpsAnalysis.Core/Data/Models/PlayerInfo.cs
@@ -42,5 +42,5 @@ public class PlayerInfo
     public int? SeasonStrength { get; internal set; }
     public int SeasonLevel { get; internal set; }
     public bool CombatState { get; internal set; }
-    public int CombatStateTime { get; internal set; }
+    public long CombatStateTime { get; internal set; }
 }
